zig fmt autogen'd code
I'm not sure where this code is generated, it seems like maybe it was at one point
but that may no longer be the case? Or I just wasn't able to find the generator.

But `zig fmt` produces these changes, which makes it a little annoying to develop
so I was hoping we could make these files `zig fmt` compliant so I can just run `zig fmt`
instead of `zig fmt && git checkout -- src/components/autogen` :)

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>