normalizer: introduce UDDC compression (48K -> 19K, better than brotli and gzip)
This change follows on to #7, and introduces a compression algorithm specifically
designed for the Decompositions.bin (implemented here) and allkeys.txt (to be
implemented later) called Unicode Data Differential Compression (UDDC).

It beats gzip and brotli compression on the `Decompositions.bin` file, and is
complementary to them (which is a nice property, as many people are likely to
use gzip or brotli compression on WebAssembly binaries that include these data
files):

| File                    | Before (bytes) | After (bytes) | Change                 |
|-------------------------|----------------|---------------|------------------------|
| `Decompositions.bin`    | 48,242         | 19,072        | -60.5% (-29,170 bytes) |
| `Decompositions.bin.br` | 24,411         | 14,783        | -39.4% (-9,628 bytes)  |
| `Decompositions.bin.gz` | 30,931         | 15,670        | -49.34% (15,261 bytes) |

* Before represents binary format without UDDC compression.
* After represents binary format with UDDC compression.
* `.br` represents `brotli <file>` compression
* `.gz` represents `gzip -9 <file>` compression

I tried to strike a balance between keeping the implementation small and straightforward,
while also compressing as much as possible. I went through several iterations and did
some analysis on the results, and then simplified the op stream produced to remove logic
that did not produce very fruitful results. This is, in my view, the best balance between
compression logic complexity/code size and return value we can achieve.

Helps #3

Signed-off-by: Stephen Gutekanst <stephen@hexops.com>