Improve display_width.center and add more test cases
This PR proposes some changes to the `center` function in `display_width`. I'm not
sure if the current behavior is intended or not, but it caused issues in my library
so I'm thus suggesting the change in the form of a PR.

First issue is that I noticed the width wasn't always honored when
centering text with padding. Status quo works such that the same amount of padding is
added to both sides, which causes the returned string to be shorter than width if
the parity of string length and width are different.

Before (output string is 8, not the requested 9):
```zig
    centered = try center(allocator, "1234", 9, "-");
    try std.testing.expectEqualSlices(u8, "--1234--", centered);
```

After:
```zig
    centered = try center(allocator, "1234", 9, "-");
    try std.testing.expectEqualSlices(u8, "--1234---", centered);
```

The second issue encountered was an error if input string was the same length as
the width. I here expected the input string to be returned. I thus added support for this.

Also added a few more test cases.